+++ src/components/fluid-container/readme.md
@@ -16,13 +16,11 @@
 
 ### Used by
 
- - [code-editor](../code-editor)
  - [nav-bar](../nav-bar)
 
 ### Graph
 ```mermaid
 graph TD;
-  code-editor --> fluid-container
   nav-bar --> fluid-container
   style fluid-container fill:#f9f,stroke:#333,stroke-width:4px
 ```
