--- src/components/req-editor/readme.md
@@ -1,37 +1,0 @@
-# code-editor
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property | Attribute | Description | Type     | Default     |
-| -------- | --------- | ----------- | -------- | ----------- |
-| `doc`    | `doc`     |             | `any`    | `'\n\n'`    |
-| `header` | `header`  |             | `string` | `undefined` |
-| `url`    | `url`     |             | `string` | `undefined` |
-
-
-## Dependencies
-
-### Used by
-
- - [code-editor](../code-editor)
-
-### Depends on
-
-- [res-editor](../res-editor)
-
-### Graph
-```mermaid
-graph TD;
-  req-editor --> res-editor
-  code-editor --> req-editor
-  style req-editor fill:#f9f,stroke:#333,stroke-width:4px
-```
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
