--- src/components/sign-in/readme.md
@@ -1,25 +1,0 @@
-# sign-in
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Dependencies
-
-### Depends on
-
-- [text-field](../text-field)
-- [plain-button](../plain-button)
-
-### Graph
-```mermaid
-graph TD;
-  sign-in --> text-field
-  sign-in --> plain-button
-  style sign-in fill:#f9f,stroke:#333,stroke-width:4px
-```
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
