+++ src/components/text-field/readme.md
@@ -15,21 +15,7 @@
 | `onClick`     | `on-click`    |             | `any`                                         | `undefined`   |
 | `placeholder` | `placeholder` |             | `string`                                      | `'something'` |
 | `type`        | `type`        |             | `"email" \| "password" \| "search" \| "text"` | `'text'`      |
-| `width`       | `width`       |             | `"auto" \| "full"`                            | `'auto'`      |
 
-
-## Dependencies
-
-### Used by
-
- - [sign-in](../sign-in)
-
-### Graph
-```mermaid
-graph TD;
-  sign-in --> text-field
-  style text-field fill:#f9f,stroke:#333,stroke-width:4px
-```
 
 ----------------------------------------------
