+++ src/index.html
@@ -10,6 +10,8 @@
     <script nomodule src="build/neptune.js"></script>
   </head>
   <body class="selection:bg-indigo-200 selection:text-violet-800">
-    <code-editor header-text="Write your Gremlin Query Here" url="http://localhost:5000/query"></code-editor>
+    <fluid-container>
+      <code-editor url="http://localhost:5000/query"></code-editor>
+    </fluid-container>
   </body>
 </html>
